refactor: style cards
