+++ src/components/cards/PreviewCard/PreviewCard.module.scss
@@ -1,4 +1,4 @@
-.preview-card {
+.previewCard {
 	display: flex;
 	flex-direction: column;
 	padding: 10px 10px;
@@ -6,17 +6,15 @@
 	max-width: 280px;
 	height: 400px;
 	width: 100%;
-	transition:
-		border 0.5s,
-		transform 0.5s;
+	transition: transform 0.5s;
 }
 
-.preview-card:hover {
+.previewCard:hover {
 	cursor: pointer;
 	transform: scale(1.05);
 }
 
-.preview-card__image-container {
+.previewCardImageContainer {
 	width: 100%;
 	height: 100%;
 	flex: 1;
@@ -27,12 +25,12 @@
 	border-radius: 20px;
 }
 
-.preview-card__image {
+.previewCardImage {
 	width: 100%;
 	height: 100%;
 }
 
-.preview-card__info {
+.previewCardInfo {
 	display: flex;
 	flex-direction: column;
 	gap: 5px;
