+++ src/components/content/HomeContent/HomeContent.module.scss
@@ -1,0 +1,7 @@
+.previewCardsContainer {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	flex-wrap: wrap;
+	gap: 20px;
+}
