+++ src/hooks/card/DetailCard/useDetailCard.tsx
