+++ src/pages/home/HomePage.module.scss
@@ -24,14 +24,6 @@
 	}
 }
 
-.previewCardsContainer {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	flex-wrap: wrap;
-	gap: 20px;
-}
-
 .detailsSection {
 	display: block;
 	position: fixed;
@@ -43,6 +35,28 @@
 	box-shadow: -2px 0 5px rgba(0, 0, 0, 0.1);
 	overflow-y: auto;
 	padding: 1rem;
+
+	button {
+		background: none;
+		border: none;
+		cursor: pointer;
+		position: absolute;
+		top: 1rem;
+		right: 1rem;
+		width: 30px;
+		height: 30px;
+		transition: transform 0.2s;
+
+		img {
+			width: 30px;
+			height: 30px;
+		}
+	}
+
+	button:hover {
+		cursor: pointer;
+		transform: scale(1.1);
+	}
 }
 
 .searchContainer {
